YH-1945: refactor filterFeatureFlag directory structure
