+++ src/features/featureFlag/filterFeatureFlag/index.ts
@@ -1,0 +1,2 @@
+export { useFeatureFlagFilters } from './model/hooks/useFeatureFlagFilters';
+export { FeatureFlagFilters } from './ui/FeatureFlagFilters/FeatureFlagFilters';
