--- src/features/featureFlag/index.ts
@@ -1,6 +1,0 @@
-import { useFeatureFlagFilters } from './model/hooks/useFeatureFlagFilters';
-import { FeatureFlagFiltersParams } from './model/types/filters';
-import { FeatureFlagFilters } from './ui/featureFlagFilters/FeatureFlagFilters';
-
-export { FeatureFlagFilters, useFeatureFlagFilters };
-export type { FeatureFlagFiltersParams };
